Add simple way of bootstrapping data
Make it easy to get test data into the system. This will save time in
the future. Right now, just enter the corresponding collection data in
the `private/bootstrap/data.json` file and they will be inserted into
the database if there are no collections there already.

If a new collection, also add a mapping from collection name to
collection instance in `collectionNameToCollection` (there's no way in
Meteor to do this that isn't super ugly).